considerable refactor of SampledAudioNode
The WebAudio API suggests that scheduled nodes can be created and forgotten each time a sound needs to be scheduled, a paradigm that works well in a garbage-collected API but not so much in high-performance C++. We change the functionality of SampledAudioNode such that it can consume a variety of scheduling events (either once, or ongoing) and internally keeps a queue of which ones should play, up to 32 voices at a time. SampledAudioNode is now more friendly to building complex sequences of sampled audio, and has no impact on the other typical use case of a single long-duration sample.