Include the affected node's name in playback state updates (#206)
Currently, the ASN_PRINT macro simply displays the state, but it doesn't give any indication as to which node is being processed. This is one of the things that confused me while trying to debug the WaveShaper example. Initially, I didn't realize there was another node that caused the scheduler to print the same messages twice (from the same thread).

The good news is that there's no problem with the same node being processed twice, after all.